Incremental commit
