+++ Source/Core/Timers/TMachine.cs
@@ -14,7 +14,6 @@
 
 using System;
 using System.Timers;
-using System.Collections.Generic;
 
 namespace Microsoft.PSharp.Timers
 {
@@ -23,67 +22,112 @@ namespace Microsoft.PSharp.Timers
 	/// </summary>
 	public abstract class TMachine : Machine
 	{
-		#region private fields
+		#region static fields
 
 		/// <summary>
-		/// Set of currently active timers.
+		/// False if machines are running in production mode.
 		/// </summary>
-		HashSet<TimerId> timers = new HashSet<TimerId>();
+		public static bool IsTestingMode;
+		#endregion
 
-        #endregion
+		#region Timer API
+		
+		/// <summary>
+		/// Start a timer. 
+		/// Depending on IsTestingMode, either a timer model is used, or the system timer is.
+		/// </summary>
+		/// <param name="IsPeriodic">Specifies whether a periodic timer is desired.</param>
+		/// <param name="period">For a periodic production timer, specifies the periodicity of the timeout events.</param>
+		/// <returns>The machine id of the created timer. Must be passed while stopping the timer.</returns>
+		protected MachineId StartTimer(bool IsPeriodic, int period)
+		{
+			// Use a model timer in testing mode
+			if (IsTestingMode)
+			{ 
+				MachineId modelTimer = this.CreateMachine(typeof(TimerModel), new InitTimer(this.Id, IsPeriodic));
+				return modelTimer;
+			}
 
-        #region Timer API
+			// In production, use a system timer
+			else
+			{
+				MachineId prodTimer = this.CreateMachine(typeof(TimerProduction), new InitTimer(this.Id, IsPeriodic, period));
+				return prodTimer;
+			}
+		}
 
-        /// <summary>
-        /// Start a timer. 
-        /// </summary>
-        /// <param name="payload">Payload of the timeout event.</param>
-        /// <param name="IsPeriodic">Specifies whether a periodic timer is desired.</param>
-        /// <param name="period">Periodicity of the timeout events.</param>
-        /// <returns>The id of the created timer.</returns>
-        protected TimerId StartTimer(object payload, bool IsPeriodic, int period)
-        {
-            // The specified period must be valid
-            this.Assert(period >= 0, "Timer period must be non-negative");
-
-            var mid = this.Runtime.CreateMachineId(this.Runtime.GetTimerMachineType());
-            var tid = new TimerId(mid, payload);
-
-            this.Runtime.CreateMachine(mid, this.Runtime.GetTimerMachineType(), new InitTimer(this.Id, tid, IsPeriodic, period));
-
-            timers.Add(tid);
-            return tid;
-
-        }
+		/// <summary>
+		/// Start a timer.
+		/// </summary>
+		/// <param name="IsPeriodic">Specifies whether a periodic timer is desired. A default period of 100ms is assumed.</param>
+		/// <returns></returns>
+		protected MachineId StartTimer(bool IsPeriodic)
+		{
+			return this.StartTimer(IsPeriodic, 100);
+		}
 
 		/// <summary>
 		/// Stop the timer.
 		/// </summary>
 		/// <param name="timer">Id of the timer machine which is being stopped.</param>
 		/// <param name="flush">Clear the queue of all timeout events generated by "timer".</param>
-		protected async void StopTimer(TimerId timer, bool flush = true)
+		protected async void StopTimer(MachineId timer, bool flush)
 		{
-			// Check if the user is indeed trying to halt a valid timer
-			this.Assert(timers.Contains(timer), "Illegal timer-id given to StopTimer");
-			timers.Remove(timer);
+			if(IsTestingMode)
+			{
+				this.Send(timer, new HaltTimer(this.Id));
+			}
+			else
+			{
+				this.Runtime.SendEvent(timer, new HaltTimer(this.Id));
+			}
 
-			this.Send(timer.mid, new HaltTimerEvent(this.Id, flush));
+			// Flush the input queue of the relevant timeout events.
+			// Send a markup event. Keep flushing eTimeout events with id==timer, until we see the markup event again.
+			// eTimeout events from timers with differed id are restored to the queue.
+			if(IsTestingMode)
+			{
+				this.Send(this.Id, new Markup());
 
-            // Flush the buffer: the timer being stopped sends a markup event to the inbox of this machine.
-            // Keep dequeuing eTimeout events (with payload being the timer being stopped), until we see the markup event.
-            if (flush)
-            {
-                while (true)
-                {
-                    var ev = await this.Receive(Tuple.Create(typeof(Markup), new Func<Event, bool>(e => true)),
-                        Tuple.Create(typeof(TimerElapsedEvent), new Func<Event, bool>(e => (e as TimerElapsedEvent).Tid == timer)));
+				while (true)
+				{
+					Event e = await this.Receive(typeof(Markup), typeof(eTimeout));
+					if(e is Markup)
+					{
+						break;
+					}
+					else if(e is eTimeout && ((e as eTimeout).id != timer))
+					{
+						this.Send(this.Id, e);
+					}
+					else
+					{
+						continue;
+					}
+				}
+			}
+			// use Runtime.SendEvent in production mode
+			else
+			{
+				this.Runtime.SendEvent(this.Id, new Markup());
 
-                    if (ev is Markup)
-                    {
-                        break;
-                    }
-                }
-            }
+				while (true)
+				{
+					Event e = await this.Receive(Tuple<typeof(Markup), true>);
+					if (e is Markup)
+					{
+						break;
+					}
+					else if (e is eTimeout && ((e as eTimeout).id != timer))
+					{
+						this.Runtime.SendEvent(this.Id, e);
+					}
+					else
+					{
+						continue;
+					}
+				}
+			}
 		}
 		#endregion
 	}
